rename file DashboardChart.tsx
