+++ src/index.ts
@@ -17,7 +17,7 @@ import FlipCard from "./components/FlipCard/FlipCard"
 import Attribution from "./components/Attributions/Attributions"
 import NextPrevSelect from "./components/NextPrevSelect/NextPrevSelect"
 import Control from "./components/Control/Control";
-import DashboardChart from "./components/DashbordChart/Dashboardchart";
+import DashboardChart from "./components/DashbordChart/DashboardChart";
 
 export { KeyFigure, DashboardElement, LoadingContainer, FlipCard, Attribution, NextPrevSelect, Control,DashboardChart }
